JWT token added and auth added

+++ controllers/tasks_controller.js
@@ -3,7 +3,9 @@ const Task = require("../models/tasks_models.js");
 // GET /tasks
 const getAllTasks = async (req, res) => {
   try {
-    const tasks = await Task.find();
+    const tasks = await Task.find({
+      user: req.user.userId
+    });
     res.status(200).json(tasks);
   } catch (error) {
     console.error(error);
@@ -14,7 +16,7 @@ const getAllTasks = async (req, res) => {
 // GET /tasks/:id
 const getTasksbyId = async (req, res) => {
   try {
-    const task = await Task.findById(req.params.id);
+    const task = await Task.findById({_id :req.params.id, user: req.user.userId});
     if (!task) return res.status(404).json({ error: "Task not found" });
     res.status(200).json(task);
   } catch (error) {
@@ -29,7 +31,7 @@ const createTask = async (req, res) => {
     const { title, description } = req.body;
     if (!title) return res.status(400).json({ error: "Title is required" });
 
-    const task = await Task.create({ title, description });
+    const task = await Task.create({ title, description , user: req.user.userId });
     res.status(201).json(task);
   } catch (error) {
     console.error(error);
@@ -40,7 +42,7 @@ const createTask = async (req, res) => {
 // PUT /tasks/:id
 const updateTask = async (req, res) => {
   try {
-    const task = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true });
+    const task = await Task.findByIdAndUpdate({ _id: req.params.id, user: req.user.userId }, req.body, { new: true });
     if (!task) return res.status(404).json({ error: "Task not found" });
     res.status(200).json(task);
   } catch (error) {
@@ -52,7 +54,7 @@ const updateTask = async (req, res) => {
 // DELETE /tasks/:id
 const deleteTask = async (req, res) => {
   try {
-    const task = await Task.findByIdAndDelete(req.params.id);
+    const task = await Task.findByIdAndDelete({ _id: req.params.id, user: req.user.userId });
     if (!task) return res.status(404).json({ error: "Task not found" });
     res.status(204).send();
   } catch (error) {
